pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chroma>=0.2.0",
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.1",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.14",
    "langchain-text-splitters>=1.1.2",
    "langchainhub>=0.1.21",
    "langgraph>=1.1.6",
    "pinecone>=8.1.2",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]
